+++ spec/commands/thredded/autofollow_users_spec.rb
@@ -69,6 +69,24 @@ module Thredded
         expect(AutofollowUsers.new(@post).new_followers).not_to include(@sara)
       end
     end
+
+    context 'with thredded_user_preferences.auto_follow_topics default true' do
+      around(:all) do |ex|
+        verbose_was = ActiveRecord::Migration.verbose
+        ActiveRecord::Migration.verbose = false
+        ActiveRecord::Migration.change_column_default :thredded_user_preferences, :auto_follow_topics, true
+        Thredded::UserPreference.reset_column_information
+        ex.run
+        ActiveRecord::Migration.change_column_default :thredded_user_preferences, :auto_follow_topics, false
+        Thredded::UserPreference.reset_column_information
+        ActiveRecord::Migration.verbose = verbose_was
+      end
+
+      it 'respects the default column value' do
+        sara = create(:user, name: 'sara', email: 'sara@example.com')
+        expect(AutofollowUsers.new(@post).new_followers).to include(sara)
+      end
+    end
   end
 
   describe AutofollowUsers, '#run' do